# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stock Investment Assistant API
  version: "1.4.0"
  description: >
    REST API for the Stock Investment Assistant backend. Provides chat, configuration, and model catalog endpoints.
    
    
    **Architecture**: Built with Flask blueprints for modular organization:
    - Core API endpoints (health, chat, config) 
    - Model management endpoints (OpenAI model operations)
    - Socket.IO integration for real-time communication
    - Conversation-aware short-term memory for multi-turn conversations
    
    
    **Changes (v1.4.0)**:
    - Full API surface reconciliation (T030): openapi.yaml now documents all 26 registered endpoints
    - Migrated chat identity field from `session_id` to `conversation_id` to align with runtime code
    - Removed `session_id` field from ChatRequest/ChatResponse schemas (not implemented in code)
    - Removed ghost endpoint GET /api/sessions/{session_id}/conversation (never implemented)
    - Removed orphaned schemas: ConversationHistory, ConversationMessage, SessionArchivedError
    - Fixed 409 Conflict error schema to match actual code behavior (`CONVERSATION_ARCHIVED` code)
    - Added Phase C management API endpoints: workspaces (5), sessions (5), conversations (5)
    - Added missing session close endpoint (POST /api/sessions/{session_id}/close)
    - Added conversation summary endpoint (GET /api/conversations/{conversation_id}/summary)
...
    - Added context_overrides to conversation create request schema
...
    
...
servers:
...
paths:
...
  /api/workspaces:
    post:
      tags: [workspace-management]
      summary: Create a new workspace
      description: >
        Creates a workspace owned by the authenticated user. The workspace starts in `active` status.
      operationId: createWorkspace
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreateRequest'
            examples:
              basic:
                value:
                  name: "Semiconductor swing trades"
                  description: "Tracking cyclical names and catalysts"
              nameOnly:
                value:
                  name: "Q4 earnings watch"
      responses:
        "201":
          description: Workspace created
          headers:
            Location:
              description: Canonical URL for the created workspace
              schema:
                type: string
                example: "/api/workspaces/abc123"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
              examples:
                created:
                  value:
                    workspace_id: "abc123"
                    name: "Semiconductor swing trades"
                    description: "Tracking cyclical names and catalysts"
                    status: "active"
                    created_at: "2026-03-23T10:00:00Z"
                    updated_at: "2026-03-23T10:00:00Z"
        "400":
          description: Validation error (missing or invalid name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                missingName:
                  value:
                    error:
                      message: "Field 'name' is required and must be a non-empty string"
                      code: "VALIDATION_ERROR"
                invalidBody:
                  value:
                    error:
                      message: "Request body must be a JSON object"
                      code: "VALIDATION_ERROR"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                serverError:
                  value:
                    error:
                      message: "Internal server error"
                      code: "INTERNAL_ERROR"
    get:
      tags: [workspace-management]
      summary: List workspaces owned by the current user
      description: >
        Returns a paginated list of workspaces owned by the authenticated user,
        sorted by `updated_at` descending.
      operationId: listWorkspaces
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return (1–100).
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Number of items to skip for pagination.
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          required: false
          description: Filter workspaces by status.
          schema:
            type: string
            enum: [active, archived]
      responses:
        "200":
          description: Paginated list of workspaces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceList'
              examples:
                twoWorkspaces:
                  value:
                    items:
                      - workspace_id: "abc123"
                        name: "Semiconductor swing trades"
                        description: "Tracking cyclical names"
                        status: "active"
                        session_count: 3
                        created_at: "2026-03-20T09:00:00Z"
                        updated_at: "2026-03-23T10:00:00Z"
                      - workspace_id: "def456"
                        name: "Q4 earnings watch"
                        description: null
                        status: "active"
                        session_count: 1
                        created_at: "2026-03-21T14:00:00Z"
                        updated_at: "2026-03-22T08:30:00Z"
                    total: 2
                    limit: 25
                    offset: 0
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
  /api/workspaces/{workspace_id}:
    get:
      tags: [workspace-management]
      summary: Get workspace detail
      description: >
        Returns the workspace with aggregate counts (session_count, active_conversation_count).
        Only the workspace owner can access it.
      operationId: getWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/XUserIdHeader'
      responses:
        "200":
          description: Workspace detail with aggregate counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceDetail'
              examples:
                detail:
                  value:
                    workspace_id: "abc123"
                    name: "Semiconductor swing trades"
                    description: "Tracking cyclical names and catalysts"
                    status: "active"
                    session_count: 3
                    active_conversation_count: 5
                    created_at: "2026-03-20T09:00:00Z"
                    updated_at: "2026-03-23T10:00:00Z"
        "403":
          description: Ownership violation — workspace belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Workspace not found"
                      code: "NOT_FOUND"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
    put:
      tags: [workspace-management]
      summary: Update a workspace
      description: >
        Updates workspace `name` and/or `description`. Only the workspace owner
        can perform this operation. The workspace must be active.
      operationId: updateWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceUpdateRequest'
            examples:
              updateName:
                value:
                  name: "Semiconductor cyclical trades"
              updateBoth:
                value:
                  name: "Semiconductor cyclical trades"
                  description: "Updated tracking for cyclical names"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceDetail'
              examples:
                updated:
                  value:
                    workspace_id: "abc123"
                    name: "Semiconductor cyclical trades"
                    description: "Updated tracking for cyclical names"
                    status: "active"
                    session_count: 3
                    active_conversation_count: 5
                    created_at: "2026-03-20T09:00:00Z"
                    updated_at: "2026-03-23T11:00:00Z"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                invalidBody:
                  value:
                    error:
                      message: "Request body must be a JSON object"
                      code: "VALIDATION_ERROR"
        "403":
          description: Ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Workspace not found"
                      code: "NOT_FOUND"
        "409":
          description: Conflict — stale entity (concurrent update detected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                conflict:
                  value:
                    error:
                      message: "Workspace was modified by another request"
                      code: "CONFLICT"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
  /api/workspaces/{workspace_id}/archive:
    post:
      tags: [workspace-management]
      summary: Archive a workspace
      description: >
        Archives the workspace and all descendant sessions and conversations atomically.
        Repeated archive requests are idempotent and leave already-archived descendants unchanged.
        Only the workspace owner can archive it.
      operationId: archiveWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/XUserIdHeader'
      responses:
        "200":
          description: Workspace archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceDetail'
              examples:
                archived:
                  value:
                    workspace_id: "abc123"
                    name: "Semiconductor swing trades"
                    description: "Tracking cyclical names and catalysts"
                    status: "archived"
                    session_count: 3
                    active_conversation_count: 0
                    created_at: "2026-03-20T09:00:00Z"
                    updated_at: "2026-03-23T12:00:00Z"
        "403":
          description: Ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Workspace not found"
                      code: "NOT_FOUND"
        "400":
          description: Validation error (e.g., invalid lifecycle transition)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                validationError:
                  value:
                    error:
                      message: "Invalid state transition"
                      code: "VALIDATION_ERROR"
        "409":
          description: Conflict — stale entity or concurrent modification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                conflict:
                  value:
                    error:
                      message: "Workspace was modified by another request"
                      code: "CONFLICT"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
...
components:
  parameters:
...
    XUserIdHeader:
      name: X-User-ID
      in: header
      required: true
      description: User identifier used for ownership scoping (Mongo ObjectId string).
      schema:
        type: string
      example: "507f1f77bcf86cd799439011"
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      description: Unique identifier of the workspace.
      schema:
        type: string
      example: "abc123"
...
  schemas:
...
    ManagementError:
      type: object
      description: >
        Error envelope used by Phase C management API endpoints. The `error` field
        is an object containing a human-readable `message`, a machine-readable `code`,
        and optional `details` and `correlation_id`.
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              description: Human-readable error description.
            code:
              type: string
              description: Machine-readable error code (e.g., VALIDATION_ERROR, NOT_FOUND, FORBIDDEN, INTERNAL_ERROR, CONFLICT).
            details:
              type: object
              additionalProperties: true
              description: Optional additional detail about the error.
            correlation_id:
              type: string
              format: uuid
              description: Optional correlation identifier for tracing.
          required: [message]
    WorkspaceCreateRequest:
      type: object
      description: Request body for creating a workspace.
      properties:
        name:
          type: string
          description: Workspace name (required, non-empty).
        description:
          type: string
          nullable: true
          description: Optional description of the workspace.
      required: [name]
    WorkspaceUpdateRequest:
      type: object
      description: Request body for updating a workspace. At least one field should be provided.
      properties:
        name:
          type: string
          description: Updated workspace name.
        description:
          type: string
          nullable: true
          description: Updated description.
    Workspace:
      type: object
      description: Base workspace resource.
      properties:
        workspace_id:
          type: string
          description: Unique identifier of the workspace.
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, archived]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [workspace_id, name, status, created_at, updated_at]
    WorkspaceDetail:
      type: object
      description: Workspace with aggregate counts for sessions and active conversations.
      allOf:
        - $ref: '#/components/schemas/Workspace'
        - type: object
          properties:
            session_count:
              type: integer
              minimum: 0
              description: Total number of sessions in this workspace.
            active_conversation_count:
              type: integer
              minimum: 0
              description: Number of active (non-archived) conversations across all sessions.
    WorkspaceListItem:
      type: object
      description: Workspace item in list responses, includes session_count.
      allOf:
        - $ref: '#/components/schemas/Workspace'
        - type: object
          properties:
            session_count:
              type: integer
              minimum: 0
              description: Total number of sessions in this workspace.
    WorkspaceList:
      type: object
      description: Paginated list of workspaces.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceListItem'
        total:
          type: integer
          minimum: 0
          description: Total number of matching workspaces.
        limit:
          type: integer
        offset:
          type: integer
      required: [items, total, limit, offset]
...
security: []